transformations :: exercise 1
with the rotate/translate reversed, we're translating the container then rotating. rotations are done around 0.0 not an arbitrary point, so by doing the translation first, we move then rotate around. Previously we did the rotation, the translation so it would appear we're rotating around a _new-point_ when we simply rotated prior to translating